Modify kadai1's report

+++ src/kadai1-2.c
@@ -2,30 +2,30 @@
 #include <math.h>
 #include <stdlib.h>
 
-double increment(double n){
+float increment(float n){
     int i;
-    double sum = 0;
+    float sum = 0;
     for (i=0; i<(int)n; i++){
-        sum += 1/((double)i+1);
+        sum += 1/((float)i+1);
     }
     return sum - log(n);
 }
 
-double decrement(double n){
+float decrement(float n){
     int i;
-    double sum = 0;
+    float sum = 0;
     for (i=(int)n; i>0; i--){
-        sum += 1/(double)i;
+        sum += 1/(float)i;
     }
     return sum - log(n);
 }
 
 int main(void){
-    double n;
+    float n;
 
     printf("nを入力して下さい。\n");
     printf("n > ");
-    scanf("%lf",&n);
+    scanf("%f",&n);
 
     printf("1/1から1/n: %f\n", increment(n));
     printf("1/nから1/1: %f\n", decrement(n));
